move everything to setup.go, delete redundant clinet methods, add cr.Status.AtProvider.LastProvisioningParameters
